feat: interactive command to add, delete and update

+++ main.py
@@ -19,6 +19,7 @@ DATABASE_URL = os.getenv('DATABASE_URL')
 TELEGRAM_API_URL = f'https://api.telegram.org/bot{TOKEN}/sendMessage'
 
 bot = telebot.TeleBot(TOKEN)
+state = {}
 
 
 def get_price_from_url(url):
@@ -55,23 +56,18 @@ def main(logger, db):
                 price = get_price_from_url(target_url)
                 logger.debug(f'Price {price} for url {target_url}')
                 if price <= target_price:
-                    logger.info(f'Hit price target for url {target_url}')
+                    logger.info(f'Hit target price {target_price} for url {target_url}')
                     send_to_chat(price, target_url, target_chat)
                     logger.debug(f'Sent message to {target_chat}')
             time.sleep(SLEEP_INTERVAL)
         except Exception as e:
           logger.error(e)
 
-def parse_message(message):
-    regex = r"(https:\/\/www\.instant-gaming\.com/.+),(\d+)"
-    url = re.findall(regex, message) 
-    if len(url) == 0:
-        return False
-    return url[0]
-
 def parse_url(message):
     regex = r"(https:\/\/www\.instant-gaming\.com/.+)"
-    url = re.findall(regex, message)  
+    url = re.findall(regex, message)
+    if len(url) < 1:
+        return False
     return url[0]
 
 @bot.message_handler(commands=["help", "start"])
@@ -79,40 +75,91 @@ def add_target(message):
     bot.reply_to(message, f"""
     /start - Start the bot.
 /help - Help message.
-/add {{URL}},{{TARGET_PRICE}} - Add a new target url to track with a price.
-/update {{URL}},{{TARGET_PRICE}} - Update an existing target url to track with a different price.
-/delete {{URL}} - Delete a tracked target.
+/add - Add a new target url to track with a price.
+/update - Update an existing target url to track with a different price.
+/delete - Delete a tracked target.
 /list - List all the tracked targets.
     """)
 
 @bot.message_handler(commands=["add"])
 def add_target(message):
+    bot.reply_to(message, f"""
+    Hi, 
+what URL do you want to track?
+    """)
+    bot.register_next_step_handler(message=message, callback=get_url_handler)
+
+def get_url_handler(message):
+    url = parse_url(message.text)
+    if not url:
+        bot.reply_to(message, f"URL must be a proper Instant Gaming URL.")
+        return
+    state[message.chat.id] = {}
+    state[message.chat.id]["url"] = url
+    bot.reply_to(message, f"""
+    Hi, 
+what target price you want to set? (for example: 35)
+    """)
+    bot.register_next_step_handler(message=message, callback=get_price_handler)
+
+def get_price_handler(message):
     try:
-        if not parse_message(message.text): 
-            bot.reply_to(message, f"URL must be a proper Instant Gaming URL. Message format must be: {{URL}},{{TARGET_PRICE}}")
-            return
-        url = parse_message(message.text)[0]
-        price = parse_message(message.text)[1]
-        db.add_target(url, float(price), message.chat.id)
+        url = state[message.chat.id]["url"]
+        price = float(message.text)
+        db.add_target(url, price, message.chat.id)
         bot.reply_to(message, f"Added target with URL {url} and price {price}")
     except ValueError:
         bot.reply_to(message, f"Target {url} already exists.")
-
-
+    except Exception as e:
+        logger.error(e)
+        bot.reply_to(message, f"Sorry, something went wrong.")
 
 @bot.message_handler(commands=["update"])
 def update_target(message):
-    if not parse_message(message.text): 
-        bot.reply_to(message, f"URL must be a proper Instant Gaming URL. Message format must be: {{URL}},{{TARGET_PRICE}}")
+    bot.reply_to(message, f"""
+    Hi, 
+what URL do you want to update?
+    """)
+    bot.register_next_step_handler(message=message, callback=get_url_for_update_handler)
+
+def get_url_for_update_handler(message):
+    url = parse_url(message.text)
+    if not url:
+        bot.reply_to(message, f"URL must be a proper Instant Gaming URL.")
         return
-    url = parse_message(message.text)[0]
-    price = parse_message(message.text)[1]
-    db.update_target_price(url, float(price), message.chat.id)
-    bot.reply_to(message, f"Updated target with URL {url} and price {price}")
+    state[message.chat.id] = {}
+    state[message.chat.id]["url"] = url
+    bot.reply_to(message, f"""
+    Hi, 
+what target price you want to set? (for example: 35)
+    """)
+    bot.register_next_step_handler(message=message, callback=get_price_for_update_handler)
+
+def get_price_for_update_handler(message):
+    try:
+        url = state[message.chat.id]["url"]
+        price = float(message.text)
+        db.update_target_price(url, price, message.chat.id)
+        bot.reply_to(message, f"Updated target with URL {url} and price {price}")
+    except ValueError:
+        bot.reply_to(message, f"Target {url} already exists.")
+    except Exception as e:
+        logger.error(e)
+        bot.reply_to(message, f"Sorry, something went wrong.")
 
 @bot.message_handler(commands=["delete"])
 def delete_target(message):
+    bot.reply_to(message, f"""
+    Hi, 
+what URL do you want to delete?
+    """)
+    bot.register_next_step_handler(message=message, callback=get_url_for_delete_handler)
+
+def get_url_for_delete_handler(message):
     url = parse_url(message.text)
+    if not url:
+        bot.reply_to(message, f"URL must be a proper Instant Gaming URL.")
+        return
     db.delete_target(url, message.chat.id)
     bot.reply_to(message, f"Target {url} deleted!")
